// client/src/components/App/Content/SignUp/SignUpForm/SignUpFields.ts
enum SignUpFields {
    MABAR_USER_NAME = 'MABAR_USER_NAME',
    FIRST_NAME = 'FIRST_NAME',
    LAST_NAME = 'LAST_NAME',
    CITY = 'CITY',
    PHONE_NUMBER = 'PHONE_NUMBER',
    ID = 'ID',
    MAIL = 'MAIL',
    OFFICE = 'OFFICE',
    DESK = 'DESK',
    LANGUAGES = 'LANGUAGES'
}

export default SignUpFields;